// apps/admin/lib/scoring/spec-types.ts
export type RuleOperator = "=" | "!=" | ">" | "<" | ">=" | "<=";
export type RuleCombinator = "AND" | "OR";